Move tests from IT to unit
